# openapi.yaml
openapi: 3.1.0
info:
  title: Trackaboo API
  version: 1.0.0
  description: |
    Financial tracking API for Trackaboo.
    All monetary amounts are float64 in base currency (no cents integer conversion).
    All dates are ISO 8601 text strings in YYYY-MM-DD format.
    All timestamps are ISO 8601 UTC strings.

servers:
  - url: /api/v1
    description: Local development

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      tags: [System]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [status, timestamp]
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time

  /metrics:
    get:
      operationId: getMetrics
      summary: Monthly aggregate metrics for the dashboard
      tags: [Metrics]
      parameters:
        - name: month
          in: query
          required: true
          description: Month in YYYY-MM format
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            example: "2026-03"
      responses:
        "200":
          description: Monthly metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyMetrics"
        "422":
          $ref: "#/components/responses/ValidationError"

  /inflows:
    get:
      operationId: listInflows
      summary: List inflow entries
      tags: [Inflows]
      parameters:
        - name: month
          in: query
          description: Filter by month (YYYY-MM)
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of inflows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InflowList"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      operationId: createInflow
      summary: Create an inflow entry
      tags: [Inflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInflowRequest"
      responses:
        "201":
          description: Inflow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inflow"
        "422":
          $ref: "#/components/responses/ValidationError"

  /inflows/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getInflow
      summary: Get a single inflow entry
      tags: [Inflows]
      responses:
        "200":
          description: Inflow entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inflow"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateInflow
      summary: Update an inflow entry
      tags: [Inflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInflowRequest"
      responses:
        "200":
          description: Updated inflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inflow"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deleteInflow
      summary: Delete an inflow entry
      tags: [Inflows]
      responses:
        "204":
          description: Deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"

  /outflows:
    get:
      operationId: listOutflows
      summary: List outflow entries
      tags: [Outflows]
      parameters:
        - name: month
          in: query
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of outflows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutflowList"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      operationId: createOutflow
      summary: Create an outflow entry
      tags: [Outflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutflowRequest"
      responses:
        "201":
          description: Outflow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outflow"
        "422":
          $ref: "#/components/responses/ValidationError"

  /outflows/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getOutflow
      summary: Get a single outflow entry
      tags: [Outflows]
      responses:
        "200":
          description: Outflow entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outflow"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateOutflow
      summary: Update an outflow entry
      tags: [Outflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOutflowRequest"
      responses:
        "200":
          description: Updated outflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outflow"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deleteOutflow
      summary: Delete an outflow entry
      tags: [Outflows]
      responses:
        "204":
          description: Deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"

  /categories:
    get:
      operationId: listCategories
      summary: List all categories
      tags: [Categories]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [outflow, inflow, any]
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      summary: Create a category
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "422":
          $ref: "#/components/responses/ValidationError"

  /categories/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: getCategory
      summary: Get a single category
      tags: [Categories]
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCategory
      summary: Update a category
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      operationId: deleteCategory
      summary: Delete a category
      description: Fails if any outflows reference this category (onDelete restrict).
      tags: [Categories]
      responses:
        "204":
          description: Deleted successfully
        "409":
          description: Cannot delete — outflows reference this category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  schemas:
    MonthlyMetrics:
      type: object
      required: [month, totalInflow, totalOutflow, netFlow, categoryBreakdown]
      properties:
        month:
          type: string
          example: "2026-03"
        totalInflow:
          type: number
          example: 5000.00
        totalOutflow:
          type: number
          example: 3200.00
        netFlow:
          type: number
          example: 1800.00
        categoryBreakdown:
          type: array
          items:
            type: object
            required: [categoryId, categoryName, amount, percentage, color]
            properties:
              categoryId:
                type: integer
              categoryName:
                type: string
              amount:
                type: number
              percentage:
                type: number
                description: "0–100, rounded to 2 decimal places"
              color:
                type: string
                description: "Hex color string, e.g. #6366f1"

    Inflow:
      type: object
      required: [id, amount, date, source, createdAt, updatedAt]
      properties:
        id:
          type: integer
        amount:
          type: number
          example: 3500.00
        date:
          type: string
          format: date
          example: "2026-03-01"
        source:
          type: string
          example: "Employer"
        description:
          type: string
          nullable: true
        categoryId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    InflowList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Inflow"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    CreateInflowRequest:
      type: object
      required: [amount, date, source]
      properties:
        amount:
          type: number
          minimum: 0.01
          example: 3500.00
        date:
          type: string
          format: date
          example: "2026-03-01"
        source:
          type: string
          minLength: 1
          maxLength: 255
          example: "Employer"
        description:
          type: string
          maxLength: 1000
          nullable: true
        categoryId:
          type: integer
          nullable: true

    UpdateInflowRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        date:
          type: string
          format: date
        source:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
          nullable: true
        categoryId:
          type: integer
          nullable: true

    Outflow:
      type: object
      required: [id, amount, date, categoryId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        amount:
          type: number
          example: 1200.00
        date:
          type: string
          format: date
          example: "2026-03-05"
        categoryId:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OutflowList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Outflow"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    CreateOutflowRequest:
      type: object
      required: [amount, date, categoryId]
      properties:
        amount:
          type: number
          minimum: 0.01
          example: 1200.00
        date:
          type: string
          format: date
          example: "2026-03-05"
        categoryId:
          type: integer
          example: 1
        description:
          type: string
          maxLength: 1000
          nullable: true

    UpdateOutflowRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        date:
          type: string
          format: date
        categoryId:
          type: integer
        description:
          type: string
          maxLength: 1000
          nullable: true

    Category:
      type: object
      required: [id, name, type, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
          example: "Groceries"
        type:
          type: string
          enum: [outflow, inflow, any]
          default: outflow
        color:
          type: string
          nullable: true
          example: "#6366f1"
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Groceries"
        type:
          type: string
          enum: [outflow, inflow, any]
          default: outflow
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          nullable: true
          example: "#6366f1"
        description:
          type: string
          maxLength: 500
          nullable: true

    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
          enum: [outflow, inflow, any]
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          nullable: true
        description:
          type: string
          maxLength: 500
          nullable: true

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items:
                  type: string
              message:
                type: string

  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"